Use Int64 [IOS]

+++ ios/Plugin/Plugin/Plugin.swift
@@ -26,33 +26,33 @@ typealias JSArray = [JSObject]
 public class DownloaderPlugin: CAPPlugin {
     static var downloads:[String:DownloadRequest] = [:]
     static var downloadsData:[String:[String:Any]] = [:]
-
+    
     @objc func initialize(){
         Alamofire.SessionManager.default.startRequestsImmediately = false;
         Alamofire.SessionManager.default.session.configuration.timeoutIntervalForRequest = 60
         Alamofire.SessionManager.default.session.configuration.timeoutIntervalForResource = 60
     }
-
+    
     @objc static func setTimeout(_ call: CAPPluginCall){
         let timeout = call.getInt("timeout") ?? 60
         Alamofire.SessionManager.default.session.configuration.timeoutIntervalForRequest = Double(timeout)
         Alamofire.SessionManager.default.session.configuration.timeoutIntervalForResource = Double(timeout)
         call.resolve()
     }
-
+    
     public override func load() {
         self.initialize()
     }
-
+    
     public func joinPath(left: String, right: String) -> String {
         let nsString: NSString = NSString.init(string:left);
         return nsString.appendingPathComponent(right);
     }
-
+    
     public func generateId() -> String{
         return NSUUID().uuidString
     }
-
+    
     @objc func createDownload(_ call: CAPPluginCall){
         let url = call.getString("url") ?? nil
         if(url == nil){
@@ -73,25 +73,25 @@ public class DownloaderPlugin: CAPPlugin {
         } else {
             fullPath = joinPath(left:tempDir.path, right:self.generateId());
         }
-
-
+        
+        
         let destination: DownloadRequest.DownloadFileDestination = { _, _ in
             let fileURL = URL(string: fullPath)
             return (fileURL!, [.removePreviousFile, .createIntermediateDirectories])
         }
-
+        
         let id = self.generateId()
-
+        
         let download = Alamofire.download(url!, to: destination)
         var task:DownloadRequest?
-        var lastRefreshTime = 0;
+        var lastRefreshTime = Int64(0);
         var lastBytesWritten =  Int64(0);
         download.downloadProgress{ progress in // called on main queue by default
             if(!progress.isFinished || !progress.isPaused){
                 var data = JSObject()
                 let currentBytes = progress.completedUnitCount
                 let totalBytes = progress.totalUnitCount
-                let currentTime = Int(Date().timeIntervalSince1970 * 1000)
+                let currentTime = Int64(Date().timeIntervalSince1970 * 1000)
                 let minTime = 100
                 var speed = Int64(0)
                 if (
@@ -102,9 +102,9 @@ public class DownloaderPlugin: CAPPlugin {
                     if (intervalTime == 0) {
                         intervalTime += 1;
                     }
-                    let updateBytes = Int(currentBytes);
+                    let updateBytes = Int64(currentBytes);
                     speed = Int64(round(Double(updateBytes / intervalTime)));
-
+                    
                     data["value"] = round(progress.fractionCompleted * 100)
                     data["currentSize"] = currentBytes
                     data["totalSize"] = totalBytes
@@ -113,12 +113,12 @@ public class DownloaderPlugin: CAPPlugin {
                     let callId = d!["call"] as! String
                     let _call = self.bridge.getSavedCall(callId)
                     _call?.success(data)
-                    lastRefreshTime = Int(Date().timeIntervalSince1970 * 1000)
+                    lastRefreshTime = Int64(Date().timeIntervalSince1970 * 1000)
                     lastBytesWritten = currentBytes ;
                 }
             }
         }
-
+        
         download
             .validate()
             .responseData(completionHandler: { (response) in
@@ -140,16 +140,16 @@ public class DownloaderPlugin: CAPPlugin {
                     break;
                 }
             })
-
+        
         task = download
-
+        
         DownloaderPlugin.downloads[id] = task
         var obj = JSObject()
         obj["value"] = id
         call.resolve(obj)
     }
     @objc func start(_ call: CAPPluginCall){
-
+        
         let id = call.getString("id") ?? nil
         if(id == nil){
             call.reject("Invalid id")
@@ -161,7 +161,7 @@ public class DownloaderPlugin: CAPPlugin {
         object["path"] = nil
         DownloaderPlugin.downloadsData[id ?? ""] = object
         let task = DownloaderPlugin.downloads[id ?? ""]
-
+        
         task?.resume()
     }
     @objc func pause(_ call: CAPPluginCall){
@@ -201,7 +201,7 @@ public class DownloaderPlugin: CAPPlugin {
             obj["value"] = nil
         }
         call.resolve(obj)
-
+        
     }
     @objc func getStatus(_ call: CAPPluginCall){
         let id = call.getString("id") ?? nil
